Split TCL-Row into own component, Add prompt to flow
